backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
psycopg2-binary
django-cors-headers
Pillow
requests
openai
djangorestframework-simplejwt
python-dotenv
spotipy
geoip2>=4.1.0
pyOpenSSL>=23.3.0
cryptography>=42.0.0
gunicorn>=21.2.0